+++ lib/random-words/array.rb
@@ -33,19 +33,39 @@ module RandomWords
     def split_names
       first_names = []
       last_names = []
+      full_names = []
       first_names_ended = false
+      last_names_ended = false
+
+      sections = [[]]
+      idx = 0
       each do |line|
-        if line.strip =~ /^[\w-]+$/ && !first_names_ended
-          first_names << line
-        elsif first_names_ended
-          last_names << line
+        if line.strip.empty? || line !~ /^[[:word:]]/i
+          idx += 1
+          sections[idx] = []
         else
-          first_names_ended = true
+          sections[idx] << line.strip
         end
       end
+
+      if sections.length == 1
+        first_names = []
+        last_names = []
+        full_names = sections[0]
+      elsif sections.length == 2
+        first_names = sections[0]
+        last_names = sections[1]
+        full_names = []
+      else
+        first_names = sections[0]
+        last_names = sections[1]
+        full_names = sections[2]
+      end
+
       first_names.delete_if(&:empty?)
       last_names.delete_if(&:empty?)
-      [first_names, last_names]
+      full_names.delete_if(&:empty?)
+      [first_names, last_names, full_names]
     end
   end
 end